Improve sync script behaviour
Specifically, this prevents some unneeded output, and also ensures that the
symlink isn't created on the learners' local drive. This was unlikely to
cause issues, but could have resulted in confusion.

It also ensures that the history is wiped at the start of each session, even
if they haven't rebuilt the container.